+++ src/app/ConnectedComponent/ConnectedComponent.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnDestroy, OnInit } from '@angular/core';
 import { Observable } from 'rxjs';
 import { BaseComponent } from '../base.component';
-import { ConnectionService } from '../connection-service.service';
+import { ConnectionService } from '../connection.service';
 import { HelloMessage, LoginMessage, Message, MessageType, WelcomeMessage } from '../Message';
 
 @Component({
