+++ src/app/connection-service.service.spec.ts
@@ -2,7 +2,7 @@ import { TestBed } from '@angular/core/testing';
 import * as rxjs  from 'rxjs';
 import * as rxws from 'rxjs/webSocket';
 
-import { ConnectionService, RXJS, Logger } from './connection-service.service';
+import { ConnectionService, RXJS, Logger } from './connection.service';
 import { ConnectedComponent } from './ConnectedComponent/ConnectedComponent.component';
 import { HelloMessage, IncomingMessage, LoginMessage, ByeMessage, Message, MessageType, LoginCredentials, WelcomeMessage } from './Message';
