+++ src/app/login-component/login.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit, Output } from '@angular/core';
 import * as rxjs from 'rxjs';
 import { ConnectedComponent } from '../ConnectedComponent/ConnectedComponent.component';
-import { ConnectionService } from '../connection-service.service';
+import { ConnectionService } from '../connection.service';
 import { LoginMessage, OutgoingMessage, WelcomeMessage, LoginCredentials, IncomingMessage, MessageType } from '../Message';
 
 @Component({
